Handle case if file is 1MB+
PyGithub poorly documents this, but the GitHub API does returns an empty contents and a "none" encoding if a file is 1MB or larger.
This was causing the AssertionError for the wrong encoding for the large files.
I've resolved it by using the "raw" download link as a backup, which supports files up to 100MB.