fix: fix github issue formatting (#471)
Attempting to fix formatting of github issues transferred to Jira.

The old format was attempting to use double-slashes (\\) to specify line breaks. This worked in the test repo but didn't look right when merged to this repo.

Now attempting to use formatted text in the yaml with |. This worked in the test repo, but I guess that's no guarantee.