bump unstructured-inference (#1140)
Pulls in fix from unstructured-inference==0.5.12:

When a pdf page doesn't have much data, it may get buffered in the write to a tempfile. If this happens, we'll hit an error reading the file back. Open to suggestions for a way to unit test this - I was creating some test files with pypdf but I couldn't trigger the error.