config: Black config requires pyproject.toml
Project-specific Black configuration like line-length cannot be
specified in setup.cfg, only the more modern pyproject.toml.

Since we probably need to move to pyproject.toml soon for building and
distribution anyway, add a `pyproject.toml` file and add our project's
Black configuration values to it. This is only line-length for now but
is likely to grow.