rfctr: move convert_and_partition_docx() up
This is a block move with no code changes to make subsequent commit
diffs easier to read.

Rationale for this location is that `convert_and_partition_docx()` is an
interface function and therefore more immediately of interest to a
reader of this module (especially a first-time reader) than the
helper callables that take part in this function's implementation.