rfctr: extract _DocxPartitioner._document
Opening the provided file with `python-docx` to produce a `Document
object` is a separate concern from processing the document contents.
Also, other methods should be free to access the document for other
purposes like accessing the document settings, etc.

Extract opening the document to a lazyproperty so when and who actually
triggers the document to be opened is of no concern and all callers get
the same instance of the document object.