docs: clearer message for sentence count skip (#1410)
Related to #744 . In the `sentence_count` function, there is a parameter
that sets a threshold for a minimum word count for a sentence to be
"counted" as a sentence. When a sentence is skipped in the count because
it doesn't meet the minimum word count, the log message is potentially
misleading as it mentions "skipping" the sentence. Without the above
context, it could be interpreted that the sentence is being skipped in
the partitioning process, which is not the case.

This PR is to reword the log message to make the situation clearer.

Co-authored-by: Ronny H <138828701+ron-unstructured@users.noreply.github.com>