chore: update CHANGELOG.md (#1420)
Move to a new CHANGELOG.md convention to more fully describe changes.
Bullets should address: what was broken? what was fixed? why does it
matter?

To assist with scanning changes, the first sentence in each bullet is in
**bold**.

Note: it's also worth looking at the rendered markdown in the branch:
https://github.com/Unstructured-IO/unstructured/blob/crag/changelog-tweak/CHANGELOG.md
rather than just the git diff.

---------

Co-authored-by: Klaijan <klaijan@unstructured.io>