typing: pptx.py type-checks clean in strict mode
As part of this I added type-hints to the three decorators used on
`partition_pptx()` which led to a few changes further afield and also
turned up a significant latent bug in the PDF partitioner.

An untyped decorator obscures the type of the function it decorates so
is a much bigger gap than just a single unhinted function.

After a painful journey I worked out that IO[bytes] is the lingua-franca
type for an open binary file and BinaryIO doesn't have the full scope
for that job. One of the benefits of changing is that
`SpooledTemporaryFile` implements IO[bytes] and so doesn't need to be
separately called out. I updated the typings/docx and pptx type-stubs to
reflect this change and applied it to partition/docx.py as well while I
was at it.