fix: don't write regex_metadata={} when no regex
Don't write an empty-dict `{}` to metadata.regex_metadata when no
regex-metadata was requested. Otherwise it will serialize (because it's
not None) when it has no meaning or is even misleading (being the same
form as no-regex-matches). Also it complicates tests that don't use
regex-meta which have to include the {} value for a field when they're
not using it.