test: add unit_utils test helpers
In particular extract the code used for JSON round-trip tests which
should all use the same logic. Having it centralized also allows us to
improve the round-tripping code if necessary in future.

The other utility gets the path for an example file which takes multiple
lines in many tests and generally reduces readability.