fix: #1461 Footer should appear at section start
It's an easy intuitive mistake to make, that a header element appears at
the beginning of a section and footers appear at the end. However a
footer is just like a header in that it applies to all pages that follow
it, until superceded by a new footer, if any.

Emit footers at the beginning of the section they're defined in, not at
the end.