fix: consolidate regex_metadata and adjust offsets
The implementation of adjusting regex-metadata match-offsets assumed
the wrong data-type so while it passed the tests, in production it
dropped all regex_metadata except that in the first section.

In fairness, this never actually happened because the overchunking fixed
in the previous commit made any element that had regex matches show up
in its own single-element chunk.

Reimplement for regex-metadata of type `Dict[str, List[RegexMetadata]]`
rather than `List[RegexMetadata]`.