fix: adjust regex_metadata offsets
The type of the regex_metadata used in the tests was different than what
appears in production. Strict type-checking caught that immediately.

The implementation of adjusting regex-metadata match-offsets assumed
that wrong data-type so while it passed the tests it wouldn't work in
production.

Reimplement for regex-metadata of type `Dict[str, List[RegexMetadata]]`
rather than `List[RegexMetadata]`.