rfctr: avoid mutation of elements
Maybe it doesn't cause a problem today, but it's poor form to mutate the
data passed in to us. Maybe one day soon it will also be used somewhere
else and we can easily avoid difficult-to-track-down bugs by simply
working on a copy.

Good chance to give a more descriptive name to the `metadata` variable
while we're at it.