xfail: add failing test for over-chunking
The sectioner (`_split_elements_by_title_and_table()`) detects semantic
unit boundaries and places those elements in separate sections. One
aspect of boundary detection is `_metadata_matches()` which returns
False when a difference in `metadata.regex_metadata` is detected between
elements.

The failure of this test demonstrates this over-chunking behavior.

The undetected introduction of this behavior can be attributed to the
blacklist vs. whitelist approach discussed in issue #1790. Testing for
this is not reliable because new metadata fields are of course not known
in advance.